Script to automatically cut rc and create PR
Includes some initial utilities for dealing with github's tokens and
API. There are number of python libraries that deal with this in a more
general way, but they all appear to be unmaintained. Since this wasn't
much code, it seemed better to control it ourselves then rely on an
unmaintained library.

JIRA: https://openedx.atlassian.net/browse/TE-1074