Task to trigger mobile app builds
This script is meant to be triggerable by a jenkins form. All it does is
propagate the jenkins environment by saving it to a json file and making
a new commit/associated branch. This will eventually kick off a CI job
to actually do the build with that configuration.

JIRA: https://openedx.atlassian.net/browse/MA-1512