Better info, error logging for PR creation
- Log URL of newly created PRs (even when couldn't tag)
- Separate errors for "couldn't create PR" and "failed to tag reviewers"
- Fix mock for pull request creation

Notes on the mocking issue:

The mocked method was returning None, which was a problem when new code
started asking the returned PR result for its `number` attribute.

The solution was to remove the `return_value` specification so that a
Mock object is returned by default, which returns something (probably
another Mock?) for the number rather than raising an exception.